I hate my life
Fixed all errors that came from merging together the branches

+++ VagtPlan.ApiService/Models/User.cs
@@ -7,8 +7,8 @@ namespace ApiService.Models
         public required string Name { get; set; }
         public string? Password { get; set; }
         public required int VactionDays { get; set; }
-        public required long DepartmentId { get; set; }
-        public required long UserRoleId { get; set; }
+        public required int DepartmentId { get; set; }
+        public required int UserRoleId { get; set; }
 
         public Department? Department { get; set; }
         public UserRole? UserRole { get; set; }
